+++ NEWS.md
@@ -1,5 +1,7 @@
 # eatGADS 0.19.1.9000
 
+* `composeVar()` for combining the information of two variables in a third variable
+
 
 # eatGADS 0.19.1